Compile where gcc was the only compiler that had ever seen this.
Two failures on the first CI run, and neither is in the port. Both are code
that has never been built by anything but gcc, because Antiphon's CI has not
run since it was written.

`BotBand.cpp` narrowed an int to KeySig's uint8_t `root` inside a braced
initialiser. gcc permits it, clang and MSVC reject it outright. It is a cast,
and the comment says why it is not decoration.

The suite needs `/Zc:preprocessor` for the reason chalkwalk-ninjam recorded
when its own suite was ported: MSVC's traditional preprocessor treats
`__VA_ARGS__` as a single token, so the argument-count dispatch picks the
wrong arm and `expect(cond, msg)` quietly becomes the comma operator.

Still 25,705 assertions.

+++ src/BotBand.cpp
@@ -4,6 +4,7 @@
 #include "BotDsp.h"
 #include "BotVoice.h"
 #include <chalkwalk/music/Euclidean.h>
+#include <cstdint>
 #include <algorithm>
 
 namespace BotBand {
@@ -254,7 +255,11 @@ chalkwalk::music::KeySig toKeySig(const MusicalKey::Key &key) {
     brightness = m::kLocrian;
     break;
   }
-  return m::KeySig{((key.tonic % 12) + 12) % 12,
+  // The cast is not decoration: `root` is a uint8_t and the expression is an
+  // int, which is a narrowing conversion in a braced initialiser. gcc allows
+  // it, clang and MSVC reject it outright, and this compiled for months inside
+  // a project whose CI never built this file with either.
+  return m::KeySig{static_cast<std::uint8_t>(((key.tonic % 12) + 12) % 12),
                    static_cast<std::int8_t>(brightness),
                    {},
                    m::ScaleType::Diatonic};
